auto_copy/requirements.txt
HandBrakeCLI
PyYAML
imdbpy
cdparanoia
abcde
lame - if you want mp3
eyed3
